Fix asset name and move assets to theme folder
